Update JUJU Content Type
Trying to develop a naming system so it’s easier to swap out on the
back end of Jekyll. May change again. Trying to keep a similar flow so
it’s mainly reposition that happens. May not work… still wrapping my
head around it. I know I could pull it off in PHP…. hmmmmm this is a
terrible description.